refactor(box-drawing): consolidate into a single file
Collapse `lib/box-drawing/{index,common,blocks,lines}.ts` into one
`lib/box-drawing.ts`. The split-file layout was over-structuring for
the size of the module — the code is one self-contained renderer for
one Unicode range, and the cross-file imports were noise.

The single file is organized in clearly-delimited sections:
  1. Common types (Weight, N/L/H/D, heavyThickness)
  2. Public API (isBoxOrBlock, drawBoxOrBlock)
  3. Block elements U+2580..U+259F (block/quadrant/fullShade helpers
     + drawBlockElement dispatch)
  4. Box-drawing lines U+2500..U+257F (drawBoxLine dispatch + EDGES
     /DASHED/ARC tables + drawEdges/drawArc/drawDashed/drawDiagonal)

No behavior change. The renderer's import path (`./box-drawing`) still
resolves correctly. The test file moved from
`lib/box-drawing/box-drawing.test.ts` to `lib/box-drawing.test.ts`
and updated its import. All 369 tests still pass.
